Get rid off compiler warnings (caomparison between signed and unsigned integer).
[[Imported from SVN: r4081]]

+++ istl/paamg/indicescoarsener.hh
@@ -239,7 +239,7 @@ namespace Dune
 
       coarseIndices.endResize();
 
-      assert(renumberer <= coarseIndices.size());
+      assert(static_cast<std::size_t>(renumberer) <= coarseIndices.size());
 
       // Reset the visited flags
       for(Iterator vertex=fineGraph.begin(); vertex != end; ++vertex)
